Move app to cmd/ and add Dockerfile
