+++ cmd/customd/Dockerfile
@@ -1,0 +1,5 @@
+FROM scratch
+ADD customd /
+ENTRYPOINT ["/customd"]
+CMD [""]
+
