+++ cmd/customd/app/examples.go
